--- a/internal/metarepos/report_collector.go
+++ b/internal/metarepos/report_collector.go
@@ -656,9 +656,8 @@ func (rce *reportCollectExecutor) processReport(response *snpb.GetReportResponse
 
 	report.Sort()
 
-	prevReport := rce.reportCtx.getReport()
-	rce.reportCtx.saveReport(report)
-	if prevReport == nil || rce.reportCtx.isExpire() {
+	prevReport, ok := rce.reportCtx.swapReport(report)
+	if !ok || rce.reportCtx.isExpire() {
 		rce.reportCtx.reload()
 		return report
 	}
@@ -746,8 +745,8 @@ func (rce *reportCollectExecutor) getClient(ctx context.Context) (reportcommitte
 }
 
 func (rce *reportCollectExecutor) getReportedVersion(lsID types.LogStreamID) (types.Version, bool) {
-	report := rce.reportCtx.getReport()
-	if report == nil {
+	report, ok := rce.reportCtx.getReport()
+	if !ok {
 		return types.InvalidVersion, false
 	}
 
@@ -922,11 +921,32 @@ func (rc *reportContext) saveReport(report *mrpb.StorageNodeUncommitReport) {
 	rc.report.UncommitReports = report.UncommitReports
 }
 
-func (rc *reportContext) getReport() *mrpb.StorageNodeUncommitReport {
+func (rc *reportContext) swapReport(newReport *mrpb.StorageNodeUncommitReport) (old mrpb.StorageNodeUncommitReport, ok bool) {
+	rc.mu.Lock()
+	defer rc.mu.Unlock()
+
+	if rc.report != nil {
+		old = *rc.report
+		ok = true
+		rc.report.Release()
+	}
+
+	rc.report = mrpb.NewStoragenodeUncommitReport(newReport.StorageNodeID)
+	rc.report.UncommitReports = newReport.UncommitReports
+
+	return old, ok
+}
+
+func (rc *reportContext) getReport() (report mrpb.StorageNodeUncommitReport, ok bool) {
 	rc.mu.RLock()
 	defer rc.mu.RUnlock()
 
-	return rc.report
+	if rc.report != nil {
+		report = *rc.report
+		ok = true
+	}
+
+	return report, ok
 }
 
 func (rc *reportContext) reload() {