@ijsong 2023-07-27
> That's something I didn't even think of. Is there any way to make [diff.UncommitReports](https://github.com/kakao/varlog/blob/519cd60f2e0bffdcba5ee4f2af4f57548fa49220/internal/metarepos/report_collector.go#L666) using pool?

Using `sync.Pool` for `diff.UncommitReports` looks promising, but the variable length makes using the pool difficult.
However, there are a few tricks you can try:
- Pooling of growable diff.UncommitReports: there will usually be an upper bound on the length.
- Use hard limit: In rare cases where the length of diff.UncommitReports can be increased than expected; we can shrink and reclaim it to the pool.
- Use reuse cycle: When the number of reusing the diff.UncommitReports is greater than the given parameter; we can discard it.

I am not convinced the above logic works well. However, some of them are being used for other open-source projects.